# add tpp_call for faster calling of non-functions