--- a/src/core/cfg.cpp
+++ b/src/core/cfg.cpp
@@ -2595,7 +2595,7 @@ CFG* computeCFG(SourceInfo* source, std::vector<AST_stmt*> body) {
         if (b->predecessors.size() == 0) {
             if (b != rtn->getStartingBlock()) {
                 rtn->print();
-                printf("%s\n", source->getName().c_str());
+                printf("%s\n", source->getName().data());
             }
             ASSERT(b == rtn->getStartingBlock(), "%d", b->idx);
         }