--- a/src/runtime/rewrite_args.h
+++ b/src/runtime/rewrite_args.h
@@ -130,6 +130,10 @@ void rearrangeArguments(ParamReceiveSpec paramspec, const ParamNames* param_name
                         Box** defaults, CallRewriteArgs* rewrite_args, bool& rewrite_success, ArgPassSpec argspec,
                         Box* arg1, Box* arg2, Box* arg3, Box** args, const std::vector<BoxedString*>* keyword_names,
                         Box*& oarg1, Box*& oarg2, Box*& oarg3, Box** oargs);
+
+// new_args should be allocated by the caller if at least three args get passed in
+ArgPassSpec bindObjIntoArgs(Box* obj, CallRewriteArgs* rewrite_args, ArgPassSpec argspec, Box*& arg1, Box*& arg2,
+                            Box*& arg3, Box** args, Box** new_args);
 } // namespace pyston
 
 #endif