Previously, we would call non-functions by doing something like `getattr(func_like_obj, '__call__')()`, ie doing an attribute lookup and then another round of figuring out how to call the object.  Now instead, we have class-level function pointers that we can set that will bypass this extra overhead.

This is "tpp_call" and not CPython's tp_call, since it uses our internal rewriter-capable interface.

There's some more cleanup that can be done; we have a similar mechanism on functions ("internal callable") which is much less useful with this new strategy.  Most of the internal callables are on classes '__call__` functions, and they should probably just be converted to tpp_call pointers.  I did this for typeCall, though it ended up being pretty messy.

The bulk of the changes in this PR are refactoring to make it easier to do all the argument munging and the associated rewrites.  This also makes the existing code cleaner, imo.  The final commit adds a tpp_call for BoxedWrapperObjects (ie what `unicode.__repr__` is), which results in a decent speedup on django_template.py:

```
       django_template.py             5.6s (2)             5.2s (2)  -7.6%
            pyxl_bench.py             3.9s (2)             4.0s (2)  +0.6%
 sqlalchemy_imperative.py             2.2s (2)             2.2s (2)  -1.1%
        django_migrate.py             1.9s (2)             1.9s (2)  -0.7%
      virtualenv_bench.py             8.0s (2)             8.1s (2)  +0.2%
                  geomean                 3.8s                 3.7s  -1.8%
```